# PERF: faster infer_dtype when numpy dtype