Performance improvement for `infer_dtype` when the array has a numpy dtype:

```python
>>> import numpy as np
>>> import pandas as pd
>>>
>>> arr_np = np.arange(1_000_000)
>>> %timeit pd.api.types.infer_dtype(arr_np)
1.54 µs ± 6.76 ns per loop  # main
196 ns ± 0.348 ns per loop # this PR
>>> arr_pd = pd.array(arr_np, dtype=pd.Int32Dtype()) 
>>> %timeit pd.api.types.infer_dtype(arr_pd)
570 ns ± 0.419 ns per loop  # both main & this PR, no difference
```

## why is this faster?

Apparantly it's faster to access `dtype.kind`than `dtype.name` for numpy dtypes:

```python
>>> %timeit arr_np.dtype.kind
125 ns ± 0.316 ns per loop
>>> %timeit arr_np.dtype.name
1.31 µs ± 7.49 ns per loop
```

So it's better to try `"kind"` before `"name"` in `lib._try_infer_map`.